Get cluster auth info from redis url
